[browser tests] [find-replace] Solidify test
Add strategic delays to avoid a race condition that happens when files are
closed and opened in quick succession.

This testcase opens and closes a '.js' file many times very quickly. The
specific file triggers the eslint and typescript-language-features plugins
to register code action handlers, that rely on the open file's to be referenced in a
structure maintained by the plugin system.

Code actions are immediately triggered upon the file being re-opened, and I think
previously registered handler do not always have enough time to be cleared, and end-up
throwing when the file can't be found in the plugin system's internal structure
(DocumentsExtImpl)

I tried but was not able to reproduce the issue manually, openning the file from
the explorer and closing it using shortcut shift-alt-w as quick as possible.

note: included a review suggestion by paul-marechal

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>