preferences: fix duplicated resolution
The `WorkspacePreferenceProvider` can do one of two things: It may
return preferences from the current workspace folder if the workspace is
a single root workspace, or it may return preferences as defined in the
workspace file if the workspace is a multi-root workspace.

Whenever we want to resolve a preference, the following scopes are
queried: `Default`, `User`, `Workspace`, and `Folder`.

When a single root workspace is opened the following happens:

- Query `Default` scope [...]
- Query `User` scope [...]
- Query `Workspace` scope: use `WorkspacePreferenceProvider` which
delegates to `FolderPreferenceProvider`.
- Query `Folder` scope: use `FoldersPreferenceProvider`.

Conclusion: The `FoldersPreferenceProvider` is queried twice.

In order to fix that issue this commit introduces a new component:
`TogglePreferenceProvider` which extends `PreferenceProvider` and wraps
another provider so that it can be enabled/disabled.

Whenever we are running a single root workspace, we'll enable the
`FoldersPreferenceProvider` used by `WorkspacePreferenceProvider` and
disable the one bound to the `Folder` scope.

Whenever we are running a multi root workspace, we'll disable the
`FoldersPreferenceProvider` used by `WorkspacePreferenceProvider` and
enable the one bound to the `Folder` scope.