playwright: improve playwright support for electron
This patch adds the possibility to disable natively rendered elements
by setting the environment variable THEIA_ELECTRON_DISABLE_NATIVE_ELEMENTS to 1.

We need this functionality for testing menu actions or use cases that
involve file choosers using electron playwright.
When the environment variable is set, the app loader enforces the
rendering of HTML menus that playwright can access.

Contributed on behalf of STMicroelectronics

Signed-off-by: Olaf Lessenich <olessenich@eclipsesource.com>