[builtins] Use vscode builtin extensions v1.70.2
Some time between 1.66.2 and 1.70.2, the references-view extension
"ms-vscode.references-view" has gone from a standalone extension,
that has its own repository, to a vscode built-in extension, that's
part of the vscode repository. With that change, the extension's
publisher has been updated from "ms-vscode" to "vscode". As of
1.70.2, this extension is referenced in the builtins extension pack
and no longer needs to be specifically referenced in `package.json`.

This PR contains the following:

Example application:
- update the builtin extension pack to v1.70.2
- remove entry for old `references-view` extension, add it to
  "theiaPluginsExcludeIds"
- remove pinning for typescript extensions.

typescript browser tests:
- use new publisher when referencing `references-view` builtin
- re-try sending "Escape" to close the peek widget if it remains opened

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>