[browser tests] update Puppeteer
The version of Puppeteer we had is quite old and includes a version of
Chromium that does not support some of the DOM API functions we use in the
code base. e.g. HTMLElement.replaceChildren() [1], that we use in
hover-service.ts

The latest version seems to work well, and supports the example above,
so I went with that. Note that Puppeteer now bundles its typing file, and
that some of the APIs have changed slightly, requiring a small adaptation
on our end.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/Element/replaceChildren#browser_compatibility

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>