[plugin-ext] Add file locking to avoid race condition writing to global plugin storage file
Fixes #8384

Added the use of a semaphore to lock access to the global plugin storage file, to avoid it
being accessed in a disorderly fashion, causing potential corruption.

As per the issue discussion, it's possible that there is also a potential race condition
caused by node's write function, not flushing its buffers to disk, but in this specific case,
it was obvious that parallel calls to set() and get() were the main cause of the problems.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>