core: refine typing of `isObject<T>`
The template parameter for `isObject` is currently used to force cast
the value passed to the function, although there is no guarantee nor
check done to ensure that this is really the case.

This commit updates the typings so that the type passed to `isObject` is
used to guide TypeScript during subsequent field type checking,
assuming each field value is `unknown` and must also be type checked.

Example:

```ts
interface MyType {
    foo: string
}

const unknownValue: unknown = { foo: 2 };

if (isObject<MyType>(unknownValue)) {
    // We enter this branch because `unknownValue` is indeed an object,
    // and TypeScript now knows we might want to access the `foo` field
    // from our interface. But we can't yet assume to know the type of
    // `foo`, so it is still typed as `unknown` and we must check:
    if (isString(unknownValue.foo)) {
        // We won't get in this branch. But if we did, then `foo` is
        // now properly typed (and checked) as `string`.
    }
}
```