Fix file watching events being lost
Fixed an issue with tracking the current `FileSystemEvents`
in the `MainFileSystemEventService`.
Before, the events were tracked in a global field which was intended to
be cleaned after an `onDidFilesChanges()` update was sent to the proxy.
This is problematic as we do not await sending the update.
Thus, the global field will be cleaned before the update is even sent
and therefore, file watching updates might get lost.
This change ensures that a field of events is kept for each update.

Note that simply awaiting the update to be sent is not enough.
This would introduce race conditions as other parallel updates might
clean the state of other updates before they are being sent.

Fixes #12260.

Contributed on behalf of STMicroelectronics