introduce `Rc<T>`
We use the `Disposable` pattern to manage most of our resources.

This pattern implies that a resource should be disposed once no longer
useful. This is easy to implement when a resource only has one owner,
but it becomes problematic once said resource gets shared/borrowed by
several other components. Then the question of "who should dispose it?"
arises.

While it is possible to design things so it is clear who should dispose
the resource, it is not always the case.

Instead we may rely on reference counting: Anyone willing to borrow a
reference to the resource may do so while incrementing a reference
counter.

This commit introduces reference counting components such as the
`ReferenceCounter` service and its `Rc` handles.