[CI] [browser tests] Fix find-replace test suite
That test suite was still experiencing intermittent failures. Playing around with it,
I found-out that there were sometimes exceptions thrown when we do not let the
language server finish its startup, before we close the editor. The LS would then start
its initialization from scratch when the editor was re-opened, increasing the chances we
would cause issues. The solution: When opening an editor, wait until the LS has finished
starting before proceeding. This causes a small delay the first time, but is instantaneous
the following times.

Also, added a new browser test file: explorer-open-close.spec.js. I suspected that there
was some flakiness in the find-replace suite, potentially caused by the mix of opening and
closing both the explorer and an editor. I was not able to find anything but with this new
test file we might, if there's any, now or later.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>