playwright: Wait for file stat nodes
The `theia-workspace.test.ts` checks for existence of file nodes in the
explorer view to verify whether the workspace has been prepared
correctly. However, occasionally it may happen that the explorer view
doesn't show them yet, as the files are just being loaded.

This resulted in occasionally flaky test executions (#12063).

To fix this problem, we now use the `existsDirectoryNode` and
`existsFileNode` checks to explicitly wait for them to appear and
only fail if the check times out.

Contributed on behalf of STMicroelectronics.
Fixes #12063

Change-Id: I4226a7d6f657c6a1363160655da4c94cd0984379