WIP - mouse focus prototype
- focus editors
- focus terminals

TODO:
- productify prototype
- add a preference for this, disabled by default
- Other widgets that need to be supported?
	- Editor tabs (stacked editors)?
	- Select active side panel container widget?
	- ???

Works with terminals moved to a secondary window, but only if there is assist
from the OS. On Ubuntu 20.04, a utility called "gnome-tweaks" can be used to
ennable this. Under "windows", then enable "Window Focus" -> "Focus on Hover".

to test/demo, use xodotool [1] to simulate keyboard input, position the mouse
over various terminals in turn,  in the Theia app, to start "top", and then
to exit it:
$ for i in {1..4}; do sleep 3; xdotool key t o p 36; done
$ for i in {1..4}; do sleep 3; xdotool key q; done

[1]: sudo apt install xdotool
ref: https://linuxhint.com/xdotool_stimulate_mouse_clicks_and_keystrokes/

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>

Inhibit terminal mouse-focus when the Quick Input is active

we should do the same for editors

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>

inhibit setting focus if code action, suggest, rename or find widget is visible

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>