core: show tabbar when items are present (#12307)
The commit fixes an issue with the `TabBarToolbar`, where at the React 18 uplift, an old `onRender` disposable in the `TabBarToolbar` class was replaced with a new-style `onRender` mechanism, which, however, doesn't appear to work reliably. Simpler is just revealing the toolbar immediately, knowing that a render will come soon.

The changes mean that on application start the toolbar will be correctly rendered without the needed to switch view containers first.