Support 'onTaskType' activation
This brings the following benefits:

1) Activates specific plugins rather than all task provider plugins.
   This is achieved by presenting the 'provided' task types to the
   user and upon selection activate only the provider(s) for that
   specific selected task type.
2) The above avoids startup delays, and reduces the memory footprint
   on the extension host by only loading the task provider plugins
   which are selected.
3) Enables the use of plugins which support 'onTaskType'
   e.g. 'Gulp', 'Grunt', 'Jake'.

NOTE: The above apply to extensions that have stopped using
activation by the command 'workbench.action.tasks.runTask' in favor
of 'onTaskType' activation. For backward compatibility the activation
by the command mentioned  above is still supported.

Additionally, Selecting the option to configure tasks
(e.g., Terminal/Configure tasks...) requires and triggers  activation
of all tasks providers.

The solution introduces two pick levels for 'provided' tasks,
one for the selection of the task type and the second one to resolve
and show the corresponding tasks to the selected type which the user
can then pick from.

The following scenarios are impacted:

1) Terminal/Run Task (for a specific task type)
	a) Long resolution (e.g., many provided tasks)
	b) quick resolution (e.g., small number of provided tasks)
2) Terminal/Configure Tasks...
3) Command palette Quick Access (e.g. ?Task)

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>