monaco-quick-input-service: set context key "inQuickOpen"
Context Key "inQuickOpen" exists in vscode and its value (true or false) reflects
whether the Quick Open, AKA Quick input, is currently open/visible or not.

This context key can be useful in Theia too, e.g. it can be used in the "when"
clause of commands registered by vscode extensions, or programmatically whenever
it's useful to know the state of the Quick Open UI.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>
Co-authored-by: FernandoAscencio <fernando.ascencio.cama@ericsson.com>