[Compatibility] Apply default to missing 'task.scope'
The 'vscode.api' for creation of tasks has two constructors, one of
them deprecated which does not require the 'scope' parameter.
Some extensions (e.g., Jake) still use the deprecated api and
therefore it is possible to receive 'Task's where 'scope' is
undefined and causes an exception in 'Theia' which prevents the use
of the related tasks.

This change fixes it by using 'TaskScope.Workspace' as a default
in such cases.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>