Associate root folder to dynamic debug configurations
The methods is the `vscode` interface `DebugConfigurationProvider`
specify the parameter `folder: WorksapceFolder | undefined'.

The parameter can be `undefined` in the context of a folderless setup
however some plugins expect it to be present always
(e.g. ms-python.python). The original implementation of dynamic debug
configurations did not provide this parameter.

This change associates the root folder to the contributed
configurations directly after the response for the call to
'provideDebugConfigurations' and it makes sure to preserve this
parameter within the pick items presented to the user.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>