Implements "secondary window" support for Electron
Fixes #11642

The main change is to prevent the secondary window from closing until
the extracted widget is removed from the window. This includes waiting
until any close handling (including dialogs) are finished.

To enable this properly, dialog support has been extended to work with
secondary windows, including support for the StylingService in secondary
windows.

Contributed on behalf of STMicroelectronics

Signed-off-by: Thomas Mäder <t.s.maeder@gmail.com>