Remove pinned version markdown-language-features
The `vscode.markdown-language-features` plugin has been pinned to
version `1.64.2` for some time, this has been a work-around to
prevent an activation problem from a missing proposed API i.e.
`onWillDropOnTextEditor`.
However the proposed API has been removed and is not longer found
in `vscode` 1.70.2 and neither 1.77.0 (next set of plugins to be
published).

This change will allow the use of the latest published version of
the plugin.

Signed-off-by: Alvaro Sanchez-Leon <alvaro.sanchez-leon@ericsson.com>