cli: Add `check:theia-extensions` command
This allows to check for duplicate versions of theia extensions only.
Added the option `only-theia-extensions` to `theia check:dependencies`.
Theia extensions are identified via their `theiaExtensions` field in the `package.json`.
The check is running recursively as long as it finds Theia extensions.
The advantage of this approach is that a shorter, but more important, list is returned.

Contributed on behalf of STMicroelectronics