debug: improve multi-root experience for configs (#12674)
The commit improves the behavior and functionality for displaying and opening debug configurations in multi-root workspaces. 

The updates mean that the list of debug configurations is properly updated when a workspace root is added/removed. The changes also update the "open configuration" command to open the `launch.json` at the root of the currently selected debug configuration.