stop reading before writing stores
The merging of on disk values using `deepmerge` is bogus.

Instead we'll completely overwrite the file. This should be fine as the
`PluginsKeyValueStorage` singleton is shared across all plugin host
processes, meaning there shouldn't be a risk for a race condition
to corrupt the file being written on disk.

Note that there is still a race condition risk when multiple Theia
backends are running on the same host, but odds should be low.