vscode: independent editor/title/run menu
In VS Code, contributions to the "editor/title/run"
menu contribution point are not added to the
ellipsis menu but to a dedicated item on the
toolbar. This commit makes Theia do the same,
except that unlike VS Code, a pop-up menu is
always presented, even if there is only one
action available.

Fixes #12687

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>