tabbar: fix command contributions regression
The merge of #12799 introduced a regression in
the rendering of MenuToolbarItems that aren't
pop-up menus but instead just commands that
happen to have a menu path: they were rendered
as pop-up menus instead of simple buttons.

This fix refines the condition for the new pop-up
menu rendering to check that the menu to be
rendered is not a command to be executed
directly on click.

Fixes #12687

Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>