fix: restore src-gen frontend production behavior (#12950)
Restores the usage of 'require' in production mode when loading
included modules.

At the moment, both in development as well as in production mode, the
frontend will be generated using the 'import' function. Webpack will
use this opportunity for code splitting, leading to the generation of
many small bundles. As they are sequentially loaded, this has a
negative effect on the startup speed of the frontend.

By using 'require' instead, weppack will produce a single bundle, which
is faster to transmit without additional roundtrips. This behavior was
used for production mode in all previous Theia versions and is now
restored.

Contributed on behalf of STMicroelectronics