Remove baseUrl from tsconfigs
According to the [tsconfig doc](https://www.typescriptlang.org/tsconfig#baseUrl):
This feature was designed for use in conjunction with AMD module loaders in the browser, and is not recommended in any other context. As of TypeScript 4.1, baseUrl is no longer required to be set when using paths.
In addition, this options does not work well with auto import resolving of Typescript-LSPs (i.e. Theia/VS Code).
Auto import might resolves to a statement starting with `src` ( e.g. ` 
import { ChannelMultiplexer } from 'src/common/message-rpc/channel';`) which does not work well with inversify and can result in runtime errors

Most of the packages don't use it anyways. This PR removes the remaining usages and adapts `src` imports.