keybindings: add context menu to keyboard shortcuts (#12791)
* keybindings: add context menu to keyboard shortcuts

Fixes the bulk of #7582.

Adds a context menu to the Keyboard Shorcuts editor
with the following menu items:

- Copy
- Copy Command ID
- Copy Command Title
- Edit Keybinding...
- Edit When Expression...
- Add Keybinding...
- Remove Keybinding
- Reset Keybinding
- Show Same Keybindings

This commit does not address the bonus part of #7582,
i.e. no keybindings are provided for the menu items.

* Address review comments

Removed excessive `public` modifiers and `setTimeout` call.

* Address review comments

Added a label for each of the new commands.