implement async fetching of extensions
This commit allows the fetching of the verified status to be performed
asynchronously during the search to avoid causing a performance hit. The
verified (or unverified) icons will pop in asynchronously when the
fetching is done.

The following has been added:
- `verified-filled` icon for verified extensions
- `verified` icon for unverified extensions (unfilled verified)
- `question` icon for extensions with an unknown verification status (appears before the fetching is completed)

A small cleanup has also been performed with the creation of the
following functions:
- `fetchVerifiedStatus`
- `updateExtensions`

Signed-off-by: Vlad Arama <vlad.arama@ericsson.com>