Basic playwright electron support (#12207)
This change allows to write tests against an Electron app by starting the
Electron app first and then attaching Playwright to it. Also, it solves
the workspace problem by creating it upfront and pointing to it on app
start.

Because playwright cannot access native UI elements, we introduce an
optional environment variable THEIA_ELECTRON_DISABLE_NATIVE_ELEMENTS
that, when set to 1, enforces the rendering of HTML-based menus and file
chooser dialogs, which can be accessed by playwright.

Playwright also requires us to disable early rendering of the window,
which we also do via an environment variable
(THEIA_ELECTRON_NO_EARLY_WINDOW).

A unified TheiaAppLoader is introduced that can be used in both browser
and electron tests. All existing tests are adjusted to use the new
AppLoader and can be run in Electron if the environment variable
USE_ELECTRON is set to true. For convenience, a yarn target
ui-tests-electron has been added to the playwright package.json.
Furthermore, the target ui-tests-ci has been added (runs all playwright
tests in the browser and in Electron) and the Github workflow has been
adjusted.

Contributed on behalf of STMicroelectronics

Signed-off-by: Olaf Lessenich <olessenich@eclipsesource.com>
Co-authored-by: Philip Langer <planger@eclipsesource.com>