Ensure that `showOpenDialog` returns correct file URI
Use `URI.file` instead of manually reconstructing the URI from the received file path in plugin ext.
This ensures that directories or files that contain a hash are not wrongly added to the fragment component of the URI and aligns it with the expected behavior of VS Code.

Fixes #13074

Contributed on behalf of ST Microelectronics