vsx-registry: implement verified extension filtering (#12995)
This commit handles the contribution of the 'verified' field when
searching for extensions in the vsx-registry. It implements a checkmark
near the publisher name when the extension's publisher is verified.
Adds a toggle button near the search bar in the vsx
registry which allows for the filtering of extensions. When toggled,
only verified extensions will be shown during the search which allows
the user to quickly filter out extensions that might be deemed unsafe.

Signed-off-by: Vlad Arama <vlad.arama@ericsson.com>